requirements.txt
openai
google-search-results
python-dotenv
streamlit
PyPDF2
python-docx
